feat(vsp): use SubpopulationDefaultNames for subpopulation attributes (#5000)
* use SubpopulationDefaultNames for subpopulation attributes

* move class

* update style
