--- contribs/vsp/src/main/java/org/matsim/contrib/vsp/scenario/SubpopulationDefaultNames.java
@@ -1,15 +1,0 @@
-package org.matsim.contrib.vsp.scenario;
-
-public class SubpopulationDefaultNames{
-	public static final String SUBPOP_COM_PERSON = "commercialPersonTraffic";
-	public static final String SUBPOP_COM_PERSON_SERVICE = "commercialPersonTraffic_service";
-	public static final String SUBPOP_GOODS = "goodsTraffic";
-	public static final String SUBPOP_PERSON = "person";
-
-
-	/**
-	 * @deprecated -- this was used early, but should not be used in newer VSP matsim scenarios.  kai, apr/26, with input from RE and KMT
-	 */
-	@Deprecated
-	public static final String SUBPOP_FREIGHT = "freight";
-}
